requirements.txt
Flask
Flask-CORS
schedule
Adafruit_DHT
RPi.GPIO
opencv-python-headless
